+++ src/runner/entrypoints/dispatch.rs
@@ -11,6 +11,7 @@ use super::super::run_distribution;
 use super::super::run_docs;
 use super::super::run_exec;
 use super::super::run_gateway;
+use super::super::run_internal_container_lease_reaper;
 use super::super::run_internal_gateway;
 use super::super::run_internal_rhai;
 use super::super::run_release;
@@ -51,6 +52,7 @@ pub(super) fn run_command_with_cwd(cmd: Command, cwd: &Path) -> Result<String, R
         Command::Tasks(args) => run_tasks(args),
         Command::InternalGateway(args) => run_internal_gateway(args),
         Command::InternalRhai(args) => run_internal_rhai(args),
+        Command::InternalContainerLeaseReaper(args) => run_internal_container_lease_reaper(args),
         Command::Task(task) => {
             crate::runner::execute::api::run_manifest_task_with_cwd(&task, cwd.to_path_buf())
         }